fix: create a run command
