+++ examples/money-transfer/compose.yaml
@@ -34,6 +34,8 @@ services:
     links:
       - server
       - temporal
+    command:
+      - run
     depends_on:
       temporal:
         condition: service_healthy
